+++ sentry_sdk/sessions.py
@@ -1,23 +1,21 @@
 import os
-import uuid
 import time
-from datetime import datetime
 from threading import Thread, Lock
 from contextlib import contextmanager
 
+import sentry_sdk
+from sentry_sdk.envelope import Envelope
+from sentry_sdk.session import Session
 from sentry_sdk._types import MYPY
 from sentry_sdk.utils import format_timestamp
 
 if MYPY:
-    import sentry_sdk
-
+    from typing import Callable
     from typing import Optional
-    from typing import Union
     from typing import Any
     from typing import Dict
+    from typing import List
     from typing import Generator
-
-    from sentry_sdk._types import SessionStatus
 
 
 def is_auto_session_tracking_enabled(hub=None):
@@ -48,38 +46,60 @@ def auto_session_tracking(hub=None):
             hub.end_session()
 
 
-def _make_uuid(
-    val,  # type: Union[str, uuid.UUID]
-):
-    # type: (...) -> uuid.UUID
-    if isinstance(val, uuid.UUID):
-        return val
-    return uuid.UUID(val)
-
-
 TERMINAL_SESSION_STATES = ("exited", "abnormal", "crashed")
+MAX_ENVELOPE_ITEMS = 100
+
+
+def make_aggregate_envelope(aggregate_states, attrs):
+    # type: (Any, Any) -> Any
+    return {"attrs": dict(attrs), "aggregates": list(aggregate_states.values())}
 
 
 class SessionFlusher(object):
     def __init__(
         self,
-        flush_func,  # type: Any
-        flush_interval=10,  # type: int
+        capture_func,  # type: Callable[[Envelope], None]
+        session_mode,  # type: str
+        flush_interval=60,  # type: int
     ):
         # type: (...) -> None
-        self.flush_func = flush_func
+        self.capture_func = capture_func
+        self.session_mode = session_mode
         self.flush_interval = flush_interval
-        self.pending = {}  # type: Dict[str, Any]
+        self.pending_sessions = []  # type: List[Any]
+        self.pending_aggregates = {}  # type: Dict[Any, Any]
         self._thread = None  # type: Optional[Thread]
         self._thread_lock = Lock()
+        self._aggregate_lock = Lock()
         self._thread_for_pid = None  # type: Optional[int]
         self._running = True
 
     def flush(self):
         # type: (...) -> None
-        pending = self.pending
-        self.pending = {}
-        self.flush_func(list(pending.values()))
+        pending_sessions = self.pending_sessions
+        self.pending_sessions = []
+
+        with self._aggregate_lock:
+            pending_aggregates = self.pending_aggregates
+            self.pending_aggregates = {}
+
+        envelope = Envelope()
+        for session in pending_sessions:
+            if len(envelope.items) == MAX_ENVELOPE_ITEMS:
+                self.capture_func(envelope)
+                envelope = Envelope()
+
+            envelope.add_session(session)
+
+        for (attrs, states) in pending_aggregates.items():
+            if len(envelope.items) == MAX_ENVELOPE_ITEMS:
+                self.capture_func(envelope)
+                envelope = Envelope()
+
+            envelope.add_sessions(make_aggregate_envelope(states, attrs))
+
+        if len(envelope.items) > 0:
+            self.capture_func(envelope)
 
     def _ensure_running(self):
         # type: (...) -> None
@@ -93,7 +113,7 @@ class SessionFlusher(object):
                 # type: (...) -> None
                 while self._running:
                     time.sleep(self.flush_interval)
-                    if self.pending and self._running:
+                    if self._running:
                         self.flush()
 
             thread = Thread(target=_thread)
@@ -103,11 +123,45 @@ class SessionFlusher(object):
             self._thread_for_pid = os.getpid()
         return None
 
+    def add_aggregate_session(
+        self, session  # type: Session
+    ):
+        # type: (...) -> None
+        # NOTE on `session.did`:
+        # the protocol can deal with buckets that have a distinct-id, however
+        # in practice we expect the python SDK to have an extremely high cardinality
+        # here, effectively making aggregation useless, therefore we do not
+        # aggregate per-did.
+
+        # For this part we can get away with using the global interpreter lock
+        with self._aggregate_lock:
+            attrs = session.get_json_attrs(with_user_info=False)
+            primary_key = tuple(sorted(attrs.items()))
+            secondary_key = session.truncated_started  # (, session.did)
+            states = self.pending_aggregates.setdefault(primary_key, {})
+            state = states.setdefault(secondary_key, {})
+
+            if "started" not in state:
+                state["started"] = format_timestamp(session.truncated_started)
+            # if session.did is not None:
+            #     state["did"] = session.did
+            if session.status == "crashed":
+                state["crashed"] = state.get("crashed", 0) + 1
+            elif session.status == "abnormal":
+                state["abnormal"] = state.get("abnormal", 0) + 1
+            elif session.errors > 0:
+                state["errored"] = state.get("errored", 0) + 1
+            else:
+                state["exited"] = state.get("exited", 0) + 1
+
     def add_session(
         self, session  # type: Session
     ):
         # type: (...) -> None
-        self.pending[session.sid.hex] = session.to_json()
+        if self.session_mode == "request":
+            self.add_aggregate_session(session)
+        else:
+            self.pending_sessions.append(session.to_json())
         self._ensure_running()
 
     def kill(self):
@@ -117,136 +171,3 @@ class SessionFlusher(object):
     def __del__(self):
         # type: (...) -> None
         self.kill()
-
-
-class Session(object):
-    def __init__(
-        self,
-        sid=None,  # type: Optional[Union[str, uuid.UUID]]
-        did=None,  # type: Optional[str]
-        timestamp=None,  # type: Optional[datetime]
-        started=None,  # type: Optional[datetime]
-        duration=None,  # type: Optional[float]
-        status=None,  # type: Optional[SessionStatus]
-        release=None,  # type: Optional[str]
-        environment=None,  # type: Optional[str]
-        user_agent=None,  # type: Optional[str]
-        ip_address=None,  # type: Optional[str]
-        errors=None,  # type: Optional[int]
-        user=None,  # type: Optional[Any]
-    ):
-        # type: (...) -> None
-        if sid is None:
-            sid = uuid.uuid4()
-        if started is None:
-            started = datetime.utcnow()
-        if status is None:
-            status = "ok"
-        self.status = status
-        self.did = None  # type: Optional[str]
-        self.started = started
-        self.release = None  # type: Optional[str]
-        self.environment = None  # type: Optional[str]
-        self.duration = None  # type: Optional[float]
-        self.user_agent = None  # type: Optional[str]
-        self.ip_address = None  # type: Optional[str]
-        self.errors = 0
-
-        self.update(
-            sid=sid,
-            did=did,
-            timestamp=timestamp,
-            duration=duration,
-            release=release,
-            environment=environment,
-            user_agent=user_agent,
-            ip_address=ip_address,
-            errors=errors,
-            user=user,
-        )
-
-    def update(
-        self,
-        sid=None,  # type: Optional[Union[str, uuid.UUID]]
-        did=None,  # type: Optional[str]
-        timestamp=None,  # type: Optional[datetime]
-        started=None,  # type: Optional[datetime]
-        duration=None,  # type: Optional[float]
-        status=None,  # type: Optional[SessionStatus]
-        release=None,  # type: Optional[str]
-        environment=None,  # type: Optional[str]
-        user_agent=None,  # type: Optional[str]
-        ip_address=None,  # type: Optional[str]
-        errors=None,  # type: Optional[int]
-        user=None,  # type: Optional[Any]
-    ):
-        # type: (...) -> None
-        # If a user is supplied we pull some data form it
-        if user:
-            if ip_address is None:
-                ip_address = user.get("ip_address")
-            if did is None:
-                did = user.get("id") or user.get("email") or user.get("username")
-
-        if sid is not None:
-            self.sid = _make_uuid(sid)
-        if did is not None:
-            self.did = str(did)
-        if timestamp is None:
-            timestamp = datetime.utcnow()
-        self.timestamp = timestamp
-        if started is not None:
-            self.started = started
-        if duration is not None:
-            self.duration = duration
-        if release is not None:
-            self.release = release
-        if environment is not None:
-            self.environment = environment
-        if ip_address is not None:
-            self.ip_address = ip_address
-        if user_agent is not None:
-            self.user_agent = user_agent
-        if errors is not None:
-            self.errors = errors
-
-        if status is not None:
-            self.status = status
-
-    def close(
-        self, status=None  # type: Optional[SessionStatus]
-    ):
-        # type: (...) -> Any
-        if status is None and self.status == "ok":
-            status = "exited"
-        if status is not None:
-            self.update(status=status)
-
-    def to_json(self):
-        # type: (...) -> Any
-        rv = {
-            "sid": str(self.sid),
-            "init": True,
-            "started": format_timestamp(self.started),
-            "timestamp": format_timestamp(self.timestamp),
-            "status": self.status,
-        }  # type: Dict[str, Any]
-        if self.errors:
-            rv["errors"] = self.errors
-        if self.did is not None:
-            rv["did"] = self.did
-        if self.duration is not None:
-            rv["duration"] = self.duration
-
-        attrs = {}
-        if self.release is not None:
-            attrs["release"] = self.release
-        if self.environment is not None:
-            attrs["environment"] = self.environment
-        if self.ip_address is not None:
-            attrs["ip_address"] = self.ip_address
-        if self.user_agent is not None:
-            attrs["user_agent"] = self.user_agent
-        if attrs:
-            rv["attrs"] = attrs
-        return rv
